pyproject.toml
[tool.poetry]
name = "AgendaBarbeiro"
version = "0.1.0"
description = ""
authors = ["Anderson"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.1"
decouple = "^0.0.7"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
markdown = "^3.5.2"
django-filter = "^23.5"
djangorestframework-simplejwt = "^5.3.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^23.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
